+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-## [7.2.0] - 2022-07-11
+## [7.2.0] - 2022-08-02
 
 ### Added
+- Add "role" command and "exists" subcommand
+  [cyberark/cyberark-conjur-cli#417](https://github.com/cyberark/cyberark-conjur-cli/pull/417)
 - Add "show" command
   [cyberark/cyberark-conjur-cli#416](https://github.com/cyberark/cyberark-conjur-cli/pull/416)
 - Support authn-ldap